Move redss_code_translators to lib
